+++ tests/02_eradiate/01_unit/experiments/test_onedim.py
@@ -132,7 +132,7 @@ def test_onedim_experiment_real_life(mode_mono):
     ctx = KernelDictContext()
 
     # Construct with typical parameters
-    test_absorption_data_set = eradiate.path_resolver.resolve(
+    test_absorption_data_set = eradiate.data.data_store.fetch(
         "tests/spectra/absorption/us76_u86_4-spectra-4000_25711.nc"
     )
     exp = OneDimExperiment(
@@ -169,7 +169,7 @@ def test_onedim_experiment_inconsistent_multiradiancemeter(mode_mono):
     ctx = KernelDictContext()
 
     # Construct with typical parameters
-    test_absorption_data_set = eradiate.path_resolver.resolve(
+    test_absorption_data_set = eradiate.data.data_store.fetch(
         "tests/spectra/absorption/us76_u86_4-spectra-4000_25711.nc"
     )
     exp = OneDimExperiment(